requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.11.12
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.11
openai==1.3.0
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.0.0
python-telegram-bot==20.7
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
